requirements.txt
Flask
Pillow
requests
gunicorn
pytz
beautifulsoup4